refactored folder structure for dashboard page categories
